Created new API

+++ edigiseva/src/main/java/com/edigiseva/repository/UserRepository.java
@@ -1,6 +1,5 @@
 package com.edigiseva.repository;
 
-import java.math.BigDecimal;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -12,6 +11,6 @@ import com.edigiseva.model.Users;
 @Repository
 public interface UserRepository extends JpaRepository<Users, Long> {
 	Optional<Users> findByEmail(String email);
-    Optional<Users> findByUuid(BigDecimal uuid);
-    Optional<Users> findByMobileNo(BigDecimal mobileNo);
+    Optional<Users> findByUuid(Long uuid);
+    Optional<Users> findByMobileNo(Long mobileNo);
 }
